+++ Shaykhullin.Network/Packet/Packet.cs
@@ -2,9 +2,9 @@ namespace Network.Core
 {
   internal class Packet : IPacket
   {
-		public int Id { get; set; }
-    public int Order { get; set; }
-    public int Length { get; set; }
+		public byte Id { get; set; }
+    public ushort Order { get; set; }
+    public byte Length { get; set; }
     public bool End { get; set; }
     public byte[] Chunk { get; set; }
 	}
